advance() past ENDMARKER is now an XHPySyntaxError
Before, this caused a nasty bug where unclosed tags would cause the
parser to loop forever in advance():

<ui:foo>
  some stuff
<ui:foo>    # NO! BAD!

I was playing around with Django at the time, so getting to the bottom
of this required a bit of detective work. top showed 100% CPU and
rapidly climbing memory for the running server process - hmm. A further
look with strace revealed that the view was being loaded right before a
pile of brk()/mremap() memory grabbing, but a pdb.set_trace() placed at
the top of the view was never reached - given that there was also a
read() on the view file, this strongly suggested XHPy was to blame.

Running the parser manually on the index.py view file confirmed this:

$ cd ~/git/xhpy
$ PYTHONPATH=. python xhpy/parser.py < index.py

and it hung there, top showing the same rampant memory consumption.
print()ing tokens in advance() revealed that it was looping on
tokenize.ENDMARKER, so I added another check: if you're trying to
advance() past ENDMARKER, you're doing it wrong, and you deserve an
XHPySyntaxError.

I also added a test case for this, based off the code that caused the
original issue.